Move security base manager
