+++ cmdb/security/auth/providers/internal_providers.py
@@ -22,7 +22,7 @@ from cmdb.security.auth.provider_config import AuthProviderConfig
 from cmdb.security.auth.provider_config_form import AuthProviderConfigForm, AuthProviderConfigFormEntry
 from cmdb.user_management import User
 from cmdb.user_management.user_manager import UserManager, UserManagerGetError
-from cmdb.utils.security import SecurityManager
+from cmdb.security.security import SecurityManager
 from cmdb.utils.system_config import SystemConfigReader
 
 LOGGER = logging.getLogger(__name__)
